+++ cycle_app/comments/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class PublicationsConfig(AppConfig):
+class CommentsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "publications"
+    name = "comments"
