+++ cycle_app/comments/models.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+class Comment(models.Model):
+    content = models.TextField()
+    reply_id = models.ForeignKey('self', null=True, blank=True, on_delete=models.CASCADE)
