+++ cycle_app/comments/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from comments.models import Comment
+
+class CommentSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Comment
+        fields = "__all__"
