--- cycle_app/publications/views.py
@@ -1,84 +1,0 @@
-from .models import Publication, Article
-from .serializers import PublicationSerializer, ArticleSerializer
-from django.http import Http404
-
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
-
-class PublicationList(APIView):
-    def get(self, request, format=None):
-        pubs = Publication.objects.all()
-        serializer = PublicationSerializer(pubs, many=True)
-        return Response(serializer.data)
-
-    def post(self, request, format=None):
-        serializer = PublicationSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-class PublicationDetail(APIView):
-    def get_object(self, pk):
-        try:
-            return Publication.objects.get(pk=pk)
-        except Publication.DoesNotExist:
-            raise Http404
-
-    def get(self, request, pk, format=None):
-        pub = self.get_object(pk)
-        serializer = PublicationSerializer(pub)
-        return Response(serializer.data)
-
-    def patch(self, request, pk, format=None):
-        pub = self.get_object(pk)
-        serializer = PublicationSerializer(pub, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk, format=None):
-        pub = self.get_object(pk)
-        pub.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-class ArticleList(APIView):
-    def get(self, request, format=None):
-        articles = Article.objects.all()
-        serializer = ArticleSerializer(articles, many=True)
-        return Response(serializer.data)
-
-    def post(self, request, format=None):
-        serializer = ArticleSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class ArticleDetail(APIView):
-    def get_object(self, pk):
-        try:
-            return Article.objects.get(pk=pk)
-        except Article.DoesNotExist:
-            raise Http404
-
-    def get(self, request, pk, format=None):
-        article = self.get_object(pk)
-        serializer = ArticleSerializer(article)
-        return Response(serializer.data)
-
-    def patch(self, request, pk, format=None):
-        article = self.get_object(pk)
-        serializer = ArticleSerializer(article, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk, format=None):
-        article = self.get_object(pk)
-        article.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
